% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompute.R
\docType{class}
\name{BaseCompute-class}
\alias{BaseCompute-class}
\alias{$,BaseCompute-method}
\alias{$<-,BaseCompute-method}
\alias{print,BaseCompute-method}
\alias{show,BaseCompute-method}
\title{BaseCompute}
\description{
This is an internal class and should not be used directly.
}
\seealso{
\link{mxComputeEM}, \link{mxComputeGradientDescent}, \link{mxComputeHessianQuality},
\link{mxComputeIterate}, \link{mxComputeNewtonRaphson}, \link{mxComputeNumericDeriv}
}
