% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxThreshold.R
\docType{class}
\name{DiscreteBase-class}
\alias{DiscreteBase-class}
\alias{DiscreteBase}
\alias{$,DiscreteBase-method}
\alias{$<-,DiscreteBase-method}
\title{An S4 base class for discrete marginal distributions}
\description{
An S4 base class for discrete marginal distributions
}
\seealso{
\link{mxMarginalPoisson}, \link{mxMarginalNegativeBinomial}
}
