\name{dzmData}
\alias{dzmData}
\docType{data}
\title{
DZ Male data
}
\description{
Data for extended twin example ETC88.R
}
\usage{data("dzmData")}
\format{
  A data frame with 1990 observations on the following 37 variables.
  \describe{
    \item{\code{famid}}{a numeric vector}
    \item{\code{e1}}{a numeric vector}
    \item{\code{e2}}{a numeric vector}
    \item{\code{e3}}{a numeric vector}
    \item{\code{e4}}{a numeric vector}
    \item{\code{e5}}{a numeric vector}
    \item{\code{e6}}{a numeric vector}
    \item{\code{e7}}{a numeric vector}
    \item{\code{e8}}{a numeric vector}
    \item{\code{e9}}{a numeric vector}
    \item{\code{e10}}{a numeric vector}
    \item{\code{e11}}{a numeric vector}
    \item{\code{e12}}{a numeric vector}
    \item{\code{e13}}{a numeric vector}
    \item{\code{e14}}{a numeric vector}
    \item{\code{e15}}{a numeric vector}
    \item{\code{e16}}{a numeric vector}
    \item{\code{e17}}{a numeric vector}
    \item{\code{e18}}{a numeric vector}
    \item{\code{a1}}{a numeric vector}
    \item{\code{a2}}{a numeric vector}
    \item{\code{a3}}{a numeric vector}
    \item{\code{a4}}{a numeric vector}
    \item{\code{a5}}{a numeric vector}
    \item{\code{a6}}{a numeric vector}
    \item{\code{a7}}{a numeric vector}
    \item{\code{a8}}{a numeric vector}
    \item{\code{a9}}{a numeric vector}
    \item{\code{a10}}{a numeric vector}
    \item{\code{a11}}{a numeric vector}
    \item{\code{a12}}{a numeric vector}
    \item{\code{a13}}{a numeric vector}
    \item{\code{a14}}{a numeric vector}
    \item{\code{a15}}{a numeric vector}
    \item{\code{a16}}{a numeric vector}
    \item{\code{a17}}{a numeric vector}
    \item{\code{a18}}{a numeric vector}
  }
}
\examples{
data(dzmData)
str(dzmData)
}
\keyword{datasets}
