% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxOptions.R
\name{imxAutoOptionValue}
\alias{imxAutoOptionValue}
\title{imxAutoOptionValue}
\usage{
imxAutoOptionValue(optionName, optionList = options()$mxOption)
}
\arguments{
\item{optionName}{Character string naming the \link{mxOption} for which a numeric or integer value is wanted.}

\item{optionList}{List of options; defaults to list of global \link{mxOption}s.
imxAutoOptionValue}
}
\description{
Convert "Auto" placeholders in global mxOptions to actual default values.
}
\details{
This is an internal function exported for documentation purposes.
Its primary purpose is to convert the on-load value of "Auto"to
valid values for \link{mxOption}s \sQuote{Gradient step size},
\sQuote{Gradient iterations}, and 
\sQuote{Function precision}--respectively, 1.0e-7, 1L, and 1e-14.
}
