% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxCheckVariables}
\alias{imxCheckVariables}
\title{imxCheckVariables}
\usage{
imxCheckVariables(flatModel, namespace)
}
\arguments{
\item{flatModel}{flatModel}

\item{namespace}{namespace}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

