% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MxConstraint.R
\docType{data}
\name{imxConstraintRelations}
\alias{imxConstraintRelations}
\title{imxConstraintRelations}
\format{\preformatted{ chr [1:3] "<" "==" ">"
}}
\usage{
imxConstraintRelations
}
\description{
A string vector of valid constraint binary relations.
}
\keyword{datasets}

