% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxCreateMatrix}
\alias{imxCreateMatrix}
\alias{imxCreateMatrix,DiagMatrix-method}
\alias{imxCreateMatrix,FullMatrix-method}
\alias{imxCreateMatrix,IdenMatrix-method}
\alias{imxCreateMatrix,LowerMatrix-method}
\alias{imxCreateMatrix,MxMatrix-method}
\alias{imxCreateMatrix,SdiagMatrix-method}
\alias{imxCreateMatrix,StandMatrix-method}
\alias{imxCreateMatrix,SymmMatrix-method}
\alias{imxCreateMatrix,UnitMatrix-method}
\alias{imxCreateMatrix,ZeroMatrix-method}
\title{Create a matrix}
\usage{
imxCreateMatrix(.Object, labels, values, free, lbound, ubound, nrow, ncol,
  byrow, name, ...)
}
\arguments{
\item{.Object}{the matrix}

\item{labels}{labels}

\item{values}{values}

\item{free}{free}

\item{lbound}{lbound}

\item{ubound}{ubound}

\item{nrow}{nrow}

\item{ncol}{ncol}

\item{byrow}{byrow}

\item{name}{name}

\item{...}{Not used.}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

