% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxBaseNamed.R
\name{imxDefaultGetSlotDisplayNames}
\alias{imxDefaultGetSlotDisplayNames}
\title{imxDefaultGetSlotDisplayNames}
\usage{
imxDefaultGetSlotDisplayNames(x, pattern = ".*")
}
\arguments{
\item{x}{The object from which to get slot names}

\item{pattern}{Initial pattern to match (default of '.*' matches any)}
}
\description{
Returns a list of display-friendly object slot names
This is an internal function exported for those people who know
what they are doing.
}
