% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxDeparse}
\alias{imxDeparse}
\alias{imxDeparse,IdenMatrix-method}
\alias{imxDeparse,MxAlgebra-method}
\alias{imxDeparse,MxConstraint-method}
\alias{imxDeparse,MxData-method}
\alias{imxDeparse,MxMatrix-method}
\alias{imxDeparse,UnitMatrix-method}
\alias{imxDeparse,ZeroMatrix-method}
\alias{imxDeparse,matrix-method}
\title{Deparse for MxObjects}
\usage{
imxDeparse(object, indent = "   ")
}
\arguments{
\item{object}{object}

\item{indent}{indent}
}
\description{
Deparse for MxObjects
}

