% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MxMultiModel.R
\name{imxDependentModels}
\alias{imxDependentModels}
\title{Are submodels dependence?}
\usage{
imxDependentModels(model)
}
\arguments{
\item{model}{model}
}
\description{
Are submodels dependence?
}

