% Generated by roxygen2 (4.0.2): do not edit by hand
\name{imxDetermineDefaultOptimizer}
\alias{imxDetermineDefaultOptimizer}
\title{imxDetermineDefaultOptimizer}
\usage{
imxDetermineDefaultOptimizer()
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
\details{
Returns a character, the default optimizer
}

