% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxDiff.R
\name{imxDiff}
\alias{imxDiff}
\title{Set difference on regular types or S4 objects}
\usage{
imxDiff(a, b, slots = c("setequal", "intersect"))
}
\arguments{
\item{a}{a}

\item{b}{b}

\item{slots}{slots}
}
\description{
Set difference on regular types or S4 objects
}
