% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxAlgebraFunctions.R
\name{imxDmvnorm}
\alias{imxDmvnorm}
\title{A C implementation of dmvnorm}
\usage{
imxDmvnorm(loc, mean, sigma)
}
\arguments{
\item{loc}{loc}

\item{mean}{mean}

\item{sigma}{sigma}
}
\description{
This API is visible to permit testing. Please do not use.
}
