% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxSummary.R
\name{imxEvalByName}
\alias{imxEvalByName}
\title{imxEvalByName}
\usage{
imxEvalByName(name, model, compute = FALSE, show = FALSE)
}
\arguments{
\item{name}{name}

\item{model}{model}

\item{compute}{compute}

\item{show}{show}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
\details{
This function should not be used in MxSummary. All summary
information should be extracted from runstate.
}
