% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MxSearchReplace.R
\name{imxExtractMethod}
\alias{imxExtractMethod}
\title{imxExtractMethod}
\usage{
imxExtractMethod(model, index)
}
\arguments{
\item{model}{model}

\item{index}{index}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

