% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxNamespace.R
\name{imxExtractReferences}
\alias{imxExtractReferences}
\title{imxExtractReferences}
\usage{
imxExtractReferences(lst)
}
\arguments{
\item{lst}{lst}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
