% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxBaseNamed.R
\name{imxExtractSlot}
\alias{imxExtractSlot}
\title{imxExtractSlot}
\usage{
imxExtractSlot(x, name)
}
\arguments{
\item{x}{The object}

\item{name}{the name of the slot}
}
\description{
Checks for and extracts a slot from the object
This is an internal function exported for those people who know
what they are doing.
}
