% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxFilterDefinitionVariables}
\alias{imxFilterDefinitionVariables}
\title{imxFilterDefinitionVariables}
\usage{
imxFilterDefinitionVariables(defVars, dataName)
}
\arguments{
\item{defVars}{defVars}

\item{dataName}{dataName}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

