% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxMultiModel.R
\name{imxFlattenModel}
\alias{imxFlattenModel}
\title{Remove heirarchical structure from model}
\usage{
imxFlattenModel(model, namespace)
}
\arguments{
\item{model}{model}

\item{namespace}{namespace}
}
\description{
Remove heirarchical structure from model
}
