% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxEval.R
\name{imxGenerateLabels}
\alias{imxGenerateLabels}
\title{imxGenerateLabels}
\usage{
imxGenerateLabels(model)
}
\arguments{
\item{model}{model}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
