% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MxNamespace.R
\name{imxGenerateNamespace}
\alias{imxGenerateNamespace}
\title{imxGenerateNamespace}
\usage{
imxGenerateNamespace(model)
}
\arguments{
\item{model}{model}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

