% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxSwift.R
\name{imxGenSwift}
\alias{imxGenSwift}
\title{imxGenSwift}
\usage{
imxGenSwift(tc, sites, sfile)
}
\arguments{
\item{tc}{tc}

\item{sites}{sites}

\item{sfile}{sfile}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
