% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxMatrix.R
\name{imxGentleResize}
\alias{imxGentleResize}
\title{Resize an MxMatrix while preserving entries}
\usage{
imxGentleResize(matrix, dimnames)
}
\arguments{
\item{matrix}{the MxMatrix to resize}

\item{dimnames}{desired dimnames for the new matrix}
}
\value{
a resized MxMatrix
}
\description{
Resize an MxMatrix while preserving entries
}
\examples{
m1 <- mxMatrix(values=1:9, nrow=3, ncol=3,
               dimnames=list(paste0('r',1:3), paste0('c',1:3)))

imxGentleResize(m1, dimnames=list(paste0('r',c(1,3,5)),
                                  paste0('c',c(2,4,6))))
}
