% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxBaseNamed.R
\name{imxGetSlotDisplayNames}
\alias{imxGetSlotDisplayNames}
\title{imxGetSlotDisplayNames}
\usage{
imxGetSlotDisplayNames(object, pattern = ".*", slotList = slotNames(object),
  showDots = FALSE, showEmpty = FALSE)
}
\arguments{
\item{object}{The object from which to get slot names}

\item{pattern}{Initial pattern to match (default of '.*' matches any)}

\item{slotList}{List of slots for which toget display names (default = slotNames(object), i.e., all)}

\item{showDots}{Include slots whose names start with '.' (default FALSE)}

\item{showEmpty}{Include slots with length-zero contents (default FALSE)}
}
\description{
Returns a list of display-friendly object slot names
This is an internal function exported for those people who know
what they are doing.
}

