% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxPath.R
\name{imxHasAlgebraOnPath}
\alias{imxHasAlgebraOnPath}
\title{imxHasAlgebraOnPath}
\usage{
imxHasAlgebraOnPath(model, submodels = TRUE, strict = FALSE)
}
\arguments{
\item{model}{an MxModel object}

\item{submodels}{logical; recursion over child models?}

\item{strict}{logical; raise error if `model` contains no paths?}
}
\description{
This is an internal function exported for those people who know
what they are doing.  This function checks if a model (or any of its
submodels) either (1) has labels on MxPaths that reference one or more
MxAlgebras, or (2) defines any of the RAM matrices as MxAlgebras.
}
