% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxConstraint.R
\name{imxHasConstraint}
\alias{imxHasConstraint}
\title{imxHasConstraint}
\usage{
imxHasConstraint(model)
}
\arguments{
\item{model}{model}
}
\description{
This is an internal function exported for those people who know
what they are doing.  This function checks if a model (or its
submodels) has at least one MxConstraint.
}
