% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxNamespace.R
\name{imxHasDefinitionVariable}
\alias{imxHasDefinitionVariable}
\title{imxHasDefinitionVariable}
\usage{
imxHasDefinitionVariable(model)
}
\arguments{
\item{model}{model}
}
\description{
This is an internal function exported for those people who know
what they are doing.  This function checks if a model (or its
submodels) has at least one definition variable.
}
