% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompute.R
\name{imxHasNPSOL}
\alias{imxHasNPSOL}
\title{imxHasNPSOL}
\usage{
imxHasNPSOL()
}
\value{
Returns TRUE if the NPSOL proprietary optimizer is compiled and
linked with OpenMx. Otherwise FALSE.
}
\description{
imxHasNPSOL
}
