% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxPenalty.R
\name{imxHasPenalty}
\alias{imxHasPenalty}
\title{imxHasPenalty}
\usage{
imxHasPenalty(model)
}
\arguments{
\item{model}{model}
}
\description{
This is an internal function exported for those people who know
what they are doing.  This function checks if a model (or its
submodels) has any penalties.
}
