% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxFitFunctionWLS.R
\name{imxHasWLS}
\alias{imxHasWLS}
\title{imxHasWLS}
\usage{
imxHasWLS(model)
}
\arguments{
\item{model}{model}
}
\description{
This is an internal function exported for those people who know
what they are doing.  This function checks if a model uses a
fitfunction with WLS units.
}
