% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxNamespace.R
\name{imxIdentifier}
\alias{imxIdentifier}
\title{imxIdentifier}
\usage{
imxIdentifier(namespace, name)
}
\arguments{
\item{namespace}{namespace}

\item{name}{name}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
