% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxMultiModel.R
\name{imxIndependentModels}
\alias{imxIndependentModels}
\title{Are submodels independent?}
\usage{
imxIndependentModels(model)
}
\arguments{
\item{model}{model}
}
\description{
Are submodels independent?
}
