% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxModel.R
\name{imxInitModel}
\alias{imxInitModel}
\alias{imxInitModel,MxLISRELModel-method}
\alias{imxInitModel,MxModel-method}
\alias{imxInitModel,MxRAMModel-method}
\title{imxInitModel}
\usage{
imxInitModel(model)
}
\arguments{
\item{model}{model}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

