% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxNamespace.R
\name{imxIsDefinitionVariable}
\alias{imxIsDefinitionVariable}
\title{imxIsDefinitionVariable}
\usage{
imxIsDefinitionVariable(name)
}
\arguments{
\item{name}{name}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
