% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxPath.R
\name{imxIsPath}
\alias{imxIsPath}
\title{imxIsPath}
\usage{
imxIsPath(value)
}
\arguments{
\item{value}{value}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
