% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxTryHard.R
\name{imxJiggle}
\alias{imxJiggle}
\title{imxJiggle}
\usage{
imxJiggle(params, lbounds, ubounds, dsn, loc, scale)
}
\arguments{
\item{params}{Numeric vector of current free parameter values.}

\item{lbounds}{Numeric vector of lower bounds on parameters.}

\item{ubounds}{Numeric vector of upper bounds on parameters.}

\item{dsn}{Character string naming the family of distribution to be used.}

\item{loc}{Numeric vector of location parameters (medians).}

\item{scale}{Numeric vector of scale parameters (see source code).}
}
\description{
Jiggle parameter values, subject to box constraints.
For internal use by mxTryHard().
This is an internal function exported for those people who know
what they are doing.
}
