% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxErrorHandling.R
\name{imxLocateFunction}
\alias{imxLocateFunction}
\title{imxLocateFunction}
\usage{
imxLocateFunction(function_name)
}
\arguments{
\item{function_name}{function_name}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
