% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxLocateIndex}
\alias{imxLocateIndex}
\title{imxLocateIndex}
\usage{
imxLocateIndex(model, name, referant)
}
\arguments{
\item{model}{model}

\item{name}{name}

\item{referant}{referant}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

