% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxNamespace.R
\name{imxLocateLabel}
\alias{imxLocateLabel}
\title{imxLocateLabel}
\usage{
imxLocateLabel(label, model, parameter)
}
\arguments{
\item{label}{label}

\item{model}{model}

\item{parameter}{parameter}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
