% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxLookupSymbolTable}
\alias{imxLookupSymbolTable}
\title{imxLookupSymbolTable}
\usage{
imxLookupSymbolTable(name)
}
\arguments{
\item{name}{name}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

