% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxModelBuilder}
\alias{imxModelBuilder}
\alias{imxModelBuilder,MxLISRELModel-method}
\alias{imxModelBuilder,MxModel-method}
\alias{imxModelBuilder,MxRAMModel-method}
\title{imxModelBuilder}
\usage{
imxModelBuilder(model, lst, name, manifestVars, latentVars, submodels, remove,
  independent)
}
\arguments{
\item{model}{model}

\item{lst}{lst}

\item{name}{name}

\item{manifestVars}{manifestVars}

\item{latentVars}{latentVars}

\item{submodels}{submodels}

\item{remove}{remove}

\item{independent}{independent}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
\details{
TODO: It probably makes sense to split this into separate
methods. For example, modelAddVariables and modelRemoveVariables
could be their own methods. This would reduce some cut&paste
duplication.
}

