% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MxTypes.R
\docType{data}
\name{imxModelTypes}
\alias{imxModelTypes}
\title{imxModelTypes}
\format{\preformatted{ list()
}}
\usage{
imxModelTypes
}
\description{
A list of supported model types
}
\keyword{datasets}

