% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxApply.R
\name{imxMpiWrap}
\alias{imxMpiWrap}
\title{imxMpiWrap}
\usage{
imxMpiWrap(fun)
}
\arguments{
\item{fun}{fun}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
