% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OriginalMx.R
\name{imxOriginalMx}
\alias{imxOriginalMx}
\title{imxOriginalMx}
\usage{
imxOriginalMx(mx.filename, output.directory)
}
\arguments{
\item{mx.filename}{mx.filename}

\item{output.directory}{output.directory}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

