% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxPenalty.R
\docType{data}
\name{imxPenaltyTypes}
\alias{imxPenaltyTypes}
\title{imxPenaltyTypes}
\format{
An object of class \code{character} of length 3.
}
\usage{
imxPenaltyTypes
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
\details{
Types of regularization penalties.
}
\keyword{datasets}
