% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxPPML.R
\name{imxPPML}
\alias{imxPPML}
\title{imxPPML}
\usage{
imxPPML(model, flag = TRUE)
}
\arguments{
\item{model}{the MxModel to evaluate}

\item{flag}{whether to potentially enable PPML}
}
\description{
Potentially enable the PPML optimization for the given model.
}
