% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxPPML.R
\name{imxPPML.Test.Test}
\alias{imxPPML.Test.Test}
\title{imxPPML.Test.Test}
\usage{
imxPPML.Test.Test(
  model,
  checkLL = TRUE,
  checkByName = FALSE,
  tolerance = 0.5,
  testEstimates = TRUE
)
}
\arguments{
\item{model}{the MxModel to evaluate}

\item{checkLL}{whether to check log likelihood}

\item{checkByName}{check values using their names}

\item{tolerance}{closeness tolerance for check}

\item{testEstimates}{whether to test for the same parameter estimates}
}
\description{
Test that PPML solutions match non-PPML solutions.
}
\details{
This is an internal function used for comparing PPML and non-PPML solutions.
Generally, non-developers will not use this function.
}
