% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxSearchReplace.R
\name{imxReplaceMethod}
\alias{imxReplaceMethod}
\title{imxReplaceMethod}
\usage{
imxReplaceMethod(x, name, value)
}
\arguments{
\item{x}{the thing}

\item{name}{name}

\item{value}{value}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
