% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxMultiModel.R
\name{imxReplaceModels}
\alias{imxReplaceModels}
\title{Replace parts of a model}
\usage{
imxReplaceModels(model, replacements)
}
\arguments{
\item{model}{model}

\item{replacements}{replacements}
}
\description{
Replace parts of a model
}
