% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxReplaceSlot}
\alias{imxReplaceSlot}
\title{imxReplaceSlot}
\usage{
imxReplaceSlot(x, name, value, check = TRUE)
}
\arguments{
\item{x}{object}

\item{name}{the name of the slot}

\item{value}{replacement value}

\item{check}{Check replacement value for validity (default TRUE)}
}
\description{
Checks for and replaces a slot from the object
This is an internal function exported for those people who know
what they are doing.
}

