% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxRun.R
\name{imxReportProgress}
\alias{imxReportProgress}
\title{Report backend progress}
\usage{
imxReportProgress(info, eraseLen)
}
\arguments{
\item{info}{the character string to print}

\item{eraseLen}{the number of characters to erase}
}
\description{
Prints a show status string to the console without emitting a
newline.
}
