% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MxReservedNames.R
\docType{data}
\name{imxReservedNames}
\alias{imxReservedNames}
\title{imxReservedNames}
\format{\preformatted{ chr [1:6] "data" "objective" "likelihood" "fitfunction" ...
}}
\usage{
imxReservedNames
}
\description{
Vector of reserved names
}
\keyword{datasets}

