% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxRobustSE.R
\name{imxRobustSE}
\alias{imxRobustSE}
\title{imxRobustSE}
\usage{
imxRobustSE(model)
}
\arguments{
\item{model}{An OpenMx model object that has been run}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
\details{
This function computes robust standard errors via a sandwich estimator.
The "bread" of the sandwich is the numerically computed inverse Hessian
of the likelihood function.  This is what is typically used for standard
errors throughout OpenMx.  The "meat" of the sandwich is the covariance
matrix of the numerically computed row derivatives of the likelihood function
(i.e. row gradients).
}
