% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxRobustSE.R
\name{imxRowGradients}
\alias{imxRowGradients}
\title{imxRowGradients}
\usage{
imxRowGradients(model, robustSE = FALSE)
}
\arguments{
\item{model}{An OpenMx model object that has been run}

\item{robustSE}{Logical; are the row gradients being requested to calculate robust standard errors?}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
\details{
This function computes the gradient for each row of data.
The returned object is a matrix with the same number of rows as the data,
and the same number of columns as there are free parameters.
}
