% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxModel.R
\name{imxSameType}
\alias{imxSameType}
\title{imxSameType}
\usage{
imxSameType(a, b)
}
\arguments{
\item{a}{a}

\item{b}{b}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
