% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MxNamespace.R
\docType{data}
\name{imxSeparatorChar}
\alias{imxSeparatorChar}
\title{imxSeparatorChar}
\format{\preformatted{ chr "."
}}
\usage{
imxSeparatorChar
}
\description{
The character between the model name and the named entity inside
the model.
}
\keyword{datasets}

