% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompute.R
\name{imxSparseInvert}
\alias{imxSparseInvert}
\title{Sparse symmetric matrix invert}
\usage{
imxSparseInvert(mat)
}
\arguments{
\item{mat}{the matrix to invert}
}
\description{
This API is visible to permit testing. Please do not use.
}
