% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxSquareMatrix}
\alias{imxSquareMatrix}
\alias{imxSquareMatrix,DiagMatrix-method}
\alias{imxSquareMatrix,IdenMatrix-method}
\alias{imxSquareMatrix,LowerMatrix-method}
\alias{imxSquareMatrix,MxMatrix-method}
\alias{imxSquareMatrix,SdiagMatrix-method}
\alias{imxSquareMatrix,StandMatrix-method}
\alias{imxSquareMatrix,SymmMatrix-method}
\title{imxSquareMatrix}
\usage{
imxSquareMatrix(.Object)
}
\arguments{
\item{.Object}{.Object}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

