% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxSymmetricMatrix}
\alias{imxSymmetricMatrix}
\alias{imxSymmetricMatrix,LowerMatrix-method}
\alias{imxSymmetricMatrix,MxMatrix-method}
\alias{imxSymmetricMatrix,SdiagMatrix-method}
\alias{imxSymmetricMatrix,StandMatrix-method}
\alias{imxSymmetricMatrix,SymmMatrix-method}
\title{imxSymmetricMatrix}
\usage{
imxSymmetricMatrix(.Object)
}
\arguments{
\item{.Object}{.Object}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

