% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxUntitled.R
\name{imxUntitledName}
\alias{imxUntitledName}
\title{imxUntitledName}
\usage{
imxUntitledName()
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
\details{
Returns a character, the name of the next untitled entity
}
