% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MxUntitled.R
\name{imxUntitledNumber}
\alias{imxUntitledNumber}
\title{imxUntitledNumber}
\usage{
imxUntitledNumber()
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
\details{
Increments the untitled number counter and returns its value
}

