% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxUntitled.R
\name{imxUntitledNumberReset}
\alias{imxUntitledNumberReset}
\title{imxUntitledNumberReset}
\usage{
imxUntitledNumberReset()
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
\details{
Resets the \link{imxUntitledNumber} counter
}
