% Generated by roxygen2 (4.0.2): do not edit by hand
\name{imxUpdateModelValues}
\alias{imxUpdateModelValues}
\title{imxUpdateModelValues}
\usage{
imxUpdateModelValues(model, flatModel, values)
}
\arguments{
\item{model}{model}

\item{flatModel}{flat model}

\item{values}{values to update}
}
\description{
Deprecated. This function does not handle parameters with equality
constraints. Do not use.
}

