% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxVerifyMatrix}
\alias{imxVerifyMatrix}
\alias{imxVerifyMatrix,DiagMatrix-method}
\alias{imxVerifyMatrix,FullMatrix-method}
\alias{imxVerifyMatrix,IdenMatrix-method}
\alias{imxVerifyMatrix,LowerMatrix-method}
\alias{imxVerifyMatrix,MxMatrix-method}
\alias{imxVerifyMatrix,SdiagMatrix-method}
\alias{imxVerifyMatrix,StandMatrix-method}
\alias{imxVerifyMatrix,SymmMatrix-method}
\alias{imxVerifyMatrix,UnitMatrix-method}
\alias{imxVerifyMatrix,ZeroMatrix-method}
\title{imxVerifyMatrix}
\usage{
imxVerifyMatrix(.Object)
}
\arguments{
\item{.Object}{.Object}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

