% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxVerifyModel}
\alias{imxVerifyModel}
\alias{imxVerifyModel,MxLISRELModel-method}
\alias{imxVerifyModel,MxModel-method}
\alias{imxVerifyModel,MxRAMModel-method}
\title{imxVerifyModel}
\usage{
imxVerifyModel(model)
}
\arguments{
\item{model}{model}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

