% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxNamespace.R
\name{imxVerifyName}
\alias{imxVerifyName}
\title{imxVerifyName}
\usage{
imxVerifyName(name, stackNumber)
}
\arguments{
\item{name}{name}

\item{stackNumber}{stackNumber}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
