% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxVerifyReference}
\alias{imxVerifyReference}
\title{imxVerifyReference}
\usage{
imxVerifyReference(reference, stackNumber)
}
\arguments{
\item{reference}{reference}

\item{stackNumber}{stackNumber}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

