% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MxAlgebraFormula-class}
\alias{MxAlgebraFormula}
\alias{MxAlgebraFormula-class}
\alias{print,MxAlgebraFormula-method}
\alias{show,MxAlgebraFormula-method}
\title{MxAlgebraFormula}
\description{
This is an internal class for the formulas used in \code{\link{mxAlgebra}} calls.
}

