% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{MxAlgebraFunction}
\alias{MxAlgebraFunction}
\alias{MxAlgebraFunction-package}
\alias{tr}
\title{Perform some function on matrices}
\usage{
tr(x)
}
\description{
There are two ways to implement algebra functions. One way is to
implement the function only in C and use omxCallAlgebra. The other
way is to implement the same function in both R and C. The
advantage of having an R implementation is that users can easily
see it by typing the function name.
}
\details{
Some OpenMx algebra functions are available in vanilla R and some
are not.  If there is already an existing R function then the C
implementation for OpenMx must match the R version with respect to
arguments and output.
}

