% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MxExpectation.R
\docType{class}
\name{MxBaseExpectation-class}
\alias{$,MxBaseExpectation-method}
\alias{$<-,MxBaseExpectation-method}
\alias{MxBaseExpectation-class}
\title{MxBaseExpectation}
\description{
The virtual base class for all expectations. Expectations contain
enough information to generate simulated data.  This is an internal
class and should not be used directly.
}
\seealso{
\link{mxExpectationNormal}, \link{mxExpectationRAM}, \link{mxExpectationLISREL}, \link{mxExpectationStateSpace},
\link{mxExpectationBA81}
}

