% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MxBaseFitFunction-class}
\alias{$,MxBaseFitFunction-method}
\alias{$<-,MxBaseFitFunction-method}
\alias{MxBaseFitFunction-class}
\title{MxBaseFitFunction}
\description{
The virtual base class for all fit functions. This is an internal class and should not be used directly.
}
\seealso{
\link{mxFitFunctionAlgebra}, \link{mxFitFunctionML}, \link{mxFitFunctionMultigroup},
\link{mxFitFunctionR}, \link{mxFitFunctionWLS}, \link{mxFitFunctionRow},
\link{mxFitFunctionGREML}
}

