% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MxBaseNamed-class}
\alias{MxBaseNamed}
\alias{MxBaseNamed-class}
\title{MxBaseNamed}
\description{
This is an internal class and should not be used directly.  It is the
base class for named entities.  Fit functions, expectations, and computes
contain this class.
}

