% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MxBaseObjectiveMetaData-class}
\alias{MxBaseObjectiveMetaData}
\alias{MxBaseObjectiveMetaData-class}
\title{MxBaseObjectiveMetaData}
\description{
This is an internal class and should not be used directly.
It is the virtual base class for all objective functions meta-data
}

