% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompare.R
\docType{class}
\name{MxCompare-class}
\alias{MxCompare-class}
\alias{$,MxCompare-method}
\alias{[,MxCompare,ANY,ANY,ANY-method}
\alias{as.data.frame.MxCompare}
\alias{print,MxCompare-method}
\alias{show,MxCompare-method}
\title{The MxCompare Class}
\description{
The MxCompare Class
}
\details{
This is an internal class structure.  You should not use it directly.
Use \code{\link{mxCompare}} instead.
}
