% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompute.R
\name{mxComputeBenchmark}
\alias{mxComputeBenchmark}
\alias{MxComputeBenchmark-class}
\title{Repeatedly invoke a series of compute objects}
\usage{
mxComputeBenchmark(steps, ..., maxIter = 500L, freeSet = NA_character_,
  maxDuration = as.numeric(NA))
}
\arguments{
\item{steps}{a list of compute objects}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{maxIter}{the maximum number of iterations}

\item{freeSet}{Names of matrices containing free variables.}

\item{maxDuration}{the maximum amount of time (in seconds) to iterate}
}
\description{
Repeatedly invoke a series of compute objects
}
