% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompute.R
\name{mxComputeDefault}
\alias{MxComputeDefault-class}
\alias{mxComputeDefault}
\title{Default compute plan}
\usage{
mxComputeDefault(freeSet = NA_character_)
}
\arguments{
\item{freeSet}{names of matrices containing free variables}
}
\description{
The default compute plan is approximately as follows:
\code{mxComputeSequence(list(mxComputeGradientDescent(),
mxComputeConfidenceInterval(), mxComputeNumericDeriv(),
mxComputeStandardError(), mxComputeReportDeriv()))}
}

