% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompute.R
\name{mxComputeGenerateData}
\alias{mxComputeGenerateData}
\alias{MxComputeGenerateData-class}
\title{Generate data}
\usage{
mxComputeGenerateData(expectation = "expectation")
}
\arguments{
\item{expectation}{a character vector of expectations to generate data for}
}
\description{
Generate data specified by the model expectations.
}
