% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompute.R
\name{mxComputeIterate}
\alias{mxComputeIterate}
\alias{MxComputeIterate-class}
\title{Repeatedly invoke a series of compute objects until change is less than tolerance}
\usage{
mxComputeIterate(
  steps,
  ...,
  maxIter = 500L,
  tolerance = 1e-09,
  verbose = 0L,
  freeSet = NA_character_,
  maxDuration = as.numeric(NA)
)
}
\arguments{
\item{steps}{a list of compute objects}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{maxIter}{the maximum number of iterations}

\item{tolerance}{iterates until maximum relative change is less than tolerance}

\item{verbose}{integer. Level of run-time diagnostic output. Set to zero to disable}

\item{freeSet}{Names of matrices containing free variables.}

\item{maxDuration}{the maximum amount of time (in seconds) to iterate}
}
\description{
One step (typically the last) must compute the fit or maxAbsChange.
}
