% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompute.R
\name{mxComputeLoadContext}
\alias{mxComputeLoadContext}
\alias{MxComputeLoadContext-class}
\title{Load contextual data to supplement checkpoint}
\usage{
mxComputeLoadContext(method = c("csv"), path = c(), column, ...,
  sep = " ")
}
\arguments{
\item{method}{name of the conduit used to load the columns.}

\item{path}{the path to the file containing the data}

\item{column}{a character vector. The column names to log.}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{sep}{the field separator character. Values on each line of the file
are separated by this character.}
}
\description{
THIS INTERFACE IS EXPERIMENTAL AND SUBJECT TO CHANGE.
}
\details{
Currently, this only supports tab separated comma separated value
format with a column header and no row names.
An \code{originalDataIsIndexOne} option is not offered. You'll need to
add an extra line at the start on your file if you wish to make
use of \code{originalDataIsIndexOne} in \code{mxComputeLoad*}.
}
\seealso{
\link{mxComputeCheckpoint}, \link{mxComputeLoadData}, \link{mxComputeLoadMatrix}
}
