% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompute.R
\name{mxComputeLoop}
\alias{mxComputeLoop}
\alias{MxComputeLoop-class}
\alias{mxComputeBenchmark}
\title{Repeatedly invoke a series of compute objects}
\usage{
mxComputeLoop(steps, ..., i = NULL, maxIter = as.integer(NA),
  freeSet = NA_character_, maxDuration = as.numeric(NA))
}
\arguments{
\item{steps}{a list of compute objects}

\item{...}{Not used.  Forces remaining arguments to be specified
by name.}

\item{i}{the values to iterate over}

\item{maxIter}{the maximum number of iterations}

\item{freeSet}{Names of matrices containing free variables.}

\item{maxDuration}{the maximum amount of time (in seconds) to
iterate}
}
\description{
When \code{i} is given then these values are iterated
        over instead of the sequence 1 to the number of
        iterations.  If \code{i} is not given then \code{maxIter}
        is set to 500 to prevent infinite loops.
}
