% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mxComputeNothing}
\alias{mxComputeNothing}
\title{Compute nothing}
\usage{
mxComputeNothing()
}
\description{
Note that this compute plan actually does nothing whereas
\code{mxComputeOnce("expectation", "nothing")} may remove the
prediction of an expectation.
}

