% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompute.R
\name{mxComputeReportExpectation}
\alias{mxComputeReportExpectation}
\alias{MxComputeReportExpectation-class}
\title{Report expectation}
\usage{
mxComputeReportExpectation(freeSet = NA_character_)
}
\arguments{
\item{freeSet}{names of matrices containing free variables}
}
\description{
Copy the internal model expectations back to R.
}
