% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompute.R
\name{mxComputeSequence}
\alias{mxComputeSequence}
\alias{MxComputeSequence-class}
\title{Invoke a series of compute objects in sequence}
\usage{
mxComputeSequence(
  steps = list(),
  ...,
  freeSet = NA_character_,
  independent = FALSE
)
}
\arguments{
\item{steps}{a list of compute objects}

\item{...}{Not used; forces argument 'freeSet' to be specified by name.}

\item{freeSet}{Names of matrices containing free parameters.}

\item{independent}{Whether the steps could be executed out-of-order.}
}
\description{
Invoke a series of compute objects in sequence
}
