% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompute.R
\name{mxComputeSetOriginalStarts}
\alias{mxComputeSetOriginalStarts}
\alias{MxComputeSetOriginalStarts-class}
\title{Reset parameter starting values}
\usage{
mxComputeSetOriginalStarts(freeSet = NA_character_)
}
\arguments{
\item{freeSet}{names of matrices containing free variables}
}
\description{
Sets the current parameter vector back to the original starting values.
}
