% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompute.R
\name{mxComputeTryCatch}
\alias{mxComputeTryCatch}
\alias{MxComputeTryCatch-class}
\title{Execute a sub-compute plan, catching errors}
\usage{
mxComputeTryCatch(plan, ..., freeSet = NA_character_)
}
\arguments{
\item{plan}{compute plan to optimize the model}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{freeSet}{names of matrices containing free variables}
}
\description{
\lifecycle{experimental}
Any error will be recorded in a subsequent checkpoint. After
execution, the context will be reset to continue computation as if
no errors has occurred.
}
\seealso{
\link{mxComputeCheckpoint}
}
