% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mxDataDynamic}
\alias{MxDataDynamic-class}
\alias{mxDataDynamic}
\alias{print,MxDataDynamic-method}
\alias{show,MxDataDynamic-method}
\title{Create dynamic data}
\usage{
mxDataDynamic(type, ..., expectation, verbose = 0L)
}
\arguments{
\item{type}{type of data}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{expectation}{the name of the expectation to provide the data}

\item{verbose}{Increase runtime debugging output}
}
\description{
Create dynamic data
}

