% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxData.R
\docType{class}
\name{MxDataStatic-class}
\alias{MxDataStatic-class}
\alias{MxDataStatic}
\alias{print,MxDataStatic-method}
\alias{show,MxDataStatic-method}
\title{Create static data}
\description{
Internal static data class.
}
\details{
Not to be used.
}
