% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MxFlatModel-class}
\alias{$,MxFlatModel-method}
\alias{$<-,MxFlatModel-method}
\alias{MxFlatModel-class}
\alias{[[,MxFlatModel-method}
\alias{[[<-,MxFlatModel-method}
\alias{names,MxFlatModel-method}
\alias{print,MxFlatModel-method}
\alias{show,MxFlatModel-method}
\title{MxFlatModel}
\description{
This is an internal class and should not be used.
}

