% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{class}
\name{MxFlatModel}
\alias{$,MxFlatModel-method}
\alias{$<-,MxFlatModel-method}
\alias{MxFlatModel}
\alias{MxFlatModel-class}
\alias{[[,MxFlatModel-method}
\alias{[[<-,MxFlatModel-method}
\alias{print,MxFlatModel-method}
\alias{show,MxFlatModel-method}
\title{MxFlatModel

This is an internal class and should not be used.}
\description{
MxFlatModel

This is an internal class and should not be used.
}

