% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxNamespace.R
\name{mxMakeNames}
\alias{mxMakeNames}
\title{mxMakeNames}
\usage{
mxMakeNames(names, unique = FALSE)
}
\arguments{
\item{names}{a character vector}

\item{unique}{whether the pass the result through \link[base]{make.unique}}
}
\description{
Adjust a character vector so that it can be used as MxMatrix column
or row names. OpenMx is (much) more restrictive than base R's make.names.
}
\examples{
demo <- c("", "103", "data", "foo.bar[3,2]", "+!", "!+")
mxMakeNames(demo, unique=TRUE)
}
\seealso{
\link[base]{make.names}
}

