% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxThreshold.R
\name{mxMarginalNegativeBinomial}
\alias{mxMarginalNegativeBinomial}
\alias{MxMarginalNegativeBinomial-class}
\alias{print,MxMarginalNegativeBinomial-method}
\alias{show,MxMarginalNegativeBinomial-method}
\alias{$,MxMarginalNegativeBinomial-method}
\alias{$<-,MxMarginalNegativeBinomial-method}
\title{Indicator with marginal Negative Binomial distribution}
\usage{
mxMarginalNegativeBinomial(
  vars,
  maxCount = NA,
  size,
  prob = c(),
  mu = c(),
  zeroInf = 0.01,
  free = TRUE,
  labels = NA,
  lbound = NA,
  ubound = NA
)
}
\arguments{
\item{vars}{character vector of manifest indicators}

\item{maxCount}{maximum observed count}

\item{size}{positive target number of successful trials}

\item{prob}{probability of success in each trial}

\item{mu}{alternative parametrization via mean}

\item{zeroInf}{zero inflation parameter in probability units}

\item{free}{logical vector indicating whether paremeters are free}

\item{labels}{character vector of parameter labels}

\item{lbound}{numeric vector of lower bounds}

\item{ubound}{numeric vector of upper bounds}
}
\value{
a list of MxMarginPoisson obects
}
\description{
Indicator with marginal Negative Binomial distribution
}
