% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxThreshold.R
\name{mxMarginalPoisson}
\alias{mxMarginalPoisson}
\alias{MxMarginalPoisson-class}
\alias{print,MxMarginalPoisson-method}
\alias{show,MxMarginalPoisson-method}
\alias{$,MxMarginalPoisson-method}
\alias{$<-,MxMarginalPoisson-method}
\title{Indicator with marginal Poisson distribution}
\usage{
mxMarginalPoisson(
  vars,
  maxCount = NA,
  lambda,
  zeroInf = 0.01,
  free = TRUE,
  labels = NA,
  lbound = 0,
  ubound = c(1, NA)
)
}
\arguments{
\item{vars}{character vector of manifest indicators}

\item{maxCount}{maximum observed count}

\item{lambda}{non-negative means}

\item{zeroInf}{zero inflation parameter in probability units}

\item{free}{logical vector indicating whether paremeters are free}

\item{labels}{character vector of parameter labels}

\item{lbound}{numeric vector of lower bounds}

\item{ubound}{numeric vector of upper bounds}
}
\value{
a list of MxMarginPoisson obects
}
\description{
Indicator with marginal Poisson distribution
}
