% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxThreshold.R
\name{mxNormalQuantiles}
\alias{mxNormalQuantiles}
\alias{omxNormalQuantiles}
\title{mxNormalQuantiles}
\usage{
mxNormalQuantiles(nBreaks, mean = 0, sd = 1)
}
\arguments{
\item{nBreaks}{the number of thresholds, or a vector of the number of thresholds}

\item{mean}{the mean of the underlying normal distribution}

\item{sd}{the standard deviation of the underlying normal distribution}
}
\value{
a vector of quantiles
}
\description{
Get quantiles from a normal distribution
}
\examples{
mxNormalQuantiles(3)
mxNormalQuantiles(3, mean=7)
mxNormalQuantiles(2, mean=1, sd=3)
}
