% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxPenalty.R
\docType{class}
\name{MxPenalty-class}
\alias{MxPenalty-class}
\alias{$,MxPenalty-method}
\alias{$<-,MxPenalty-method}
\alias{[[<-,MxPenalty-method}
\title{MxPenalty}
\description{
This is an internal class and should not be used directly.
}
