% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxRun.R
\name{mxPenaltySearch}
\alias{mxPenaltySearch}
\title{mxPenaltySearch}
\usage{
mxPenaltySearch(
  model,
  ...,
  silent = FALSE,
  suppressWarnings = FALSE,
  unsafe = FALSE,
  checkpoint = FALSE,
  useSocket = FALSE,
  onlyFrontend = FALSE,
  beginMessage = !silent
)
}
\arguments{
\item{model}{A \link{MxModel} object to be optimized.}

\item{...}{Not used.  Forces remaining arguments to be specified by name}

\item{silent}{A boolean indicating whether to print status to terminal.}

\item{suppressWarnings}{A boolean indicating whether to suppress warnings.}

\item{unsafe}{A boolean indicating whether to ignore errors.}

\item{checkpoint}{A boolean indicating whether to periodically write parameter values to a file.}

\item{useSocket}{A boolean indicating whether to periodically write parameter values to a socket.}

\item{onlyFrontend}{A boolean indicating whether to run only front-end model transformations.}

\item{beginMessage}{A boolean indicating whether to print the number of parameters before invoking the backend.}
}
\description{
Grid search for the best \link[=MxPenalty-class]{MxPenalty} hyperparameters. Uses
\link{omxDefaultComputePlan} with
\code{penaltySearch=TRUE}. Specifically, wraps
\link{mxComputeGradientDescent} with \link{mxComputePenaltySearch}
and passes the model to the backend.
}
