% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxPenalty.R
\name{mxPenaltyZap}
\alias{mxPenaltyZap}
\title{mxPenaltyZap}
\usage{
mxPenaltyZap(model, silent = FALSE)
}
\arguments{
\item{model}{an \link{MxModel}}

\item{silent}{whether to suppress diagnostic output}
}
\description{
Fix any free parameters within \code{epsilon} of zero to
zero. These parameters are no longer estimated. Remove all
\link[=MxPenalty-class]{MxPenalty} objects from the model.  This is envisioned to be
used after using \link{mxPenaltySearch} to locate the best penalty
hyperparameters and apply penalties to model estimation. While
penalties can simplify a model, they also bias parameters toward
zero. By re-estimating the model after using \code{mxPenaltyZap},
parameters that remain free are likely to exhibit less bias.
}
