% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxGraph.R
\docType{class}
\name{MxRAMGraph-class}
\alias{MxRAMGraph-class}
\alias{MxRAMGraph}
\title{MxRAMGraph}
\description{
This is an internal class and should not be used directly.
It is a class for RAM directed graphs.
}
