% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxRAMModel.R
\docType{class}
\name{MxRAMModel-class}
\alias{MxRAMModel-class}
\alias{$<-,MxRAMModel-method}
\alias{[[<-,MxRAMModel-method}
\title{MxRAMModel}
\description{
This is an internal class and should not be used directly.
}
