% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxRetro.R
\name{mxRetro}
\alias{mxRetro}
\title{Return random classic Mx error message}
\usage{
mxRetro()
}
\description{
This function allows you to obtain a classic Mx error message.
The message returned is random.
}
\details{
If you're a nostalgic old sod and you miss the warm, fuzzy
feelings you got from reading one of Mike Neale's patented error messages,
then this function is here to save you from the depths of dire depression.
All credit for this function is due to Sarah Medland, but of course the wisdom
is from Mike Neale.
}
\examples{
require(OpenMx)
mxRetro()
}
\references{
- \url{https://en.wikipedia.org/wiki/OpenMx}
}
\seealso{
- \code{\link{omxBrownie}}
}
