% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxExpectationBA81.R
\name{mxSimplify2Array}
\alias{mxSimplify2Array}
\title{Like simplify2array but works with vectors of different lengths}
\usage{
mxSimplify2Array(x, higher = FALSE)
}
\arguments{
\item{x}{a list of vectors}

\item{higher}{whether to produce a higher rank array (defaults to FALSE)}
}
\description{
Vectors are filled column-by-column into a matrix. Shorter vectors
are padded with NAs to fill whole columns.
}
\examples{
v1 <- 1:3
v2 <- 4:5
v3 <- 6:10
mxSimplify2Array(list(v1,v2,v3))

#     [,1] [,2] [,3]
# [1,]    1    4    6
# [2,]    2    5    7
# [3,]    3   NA    8
# [4,]   NA   NA    9
# [5,]   NA   NA   10
}
