\name{nhanesDemo}
\alias{nhanesDemo}
\docType{data}
\title{
Modified National Health and Nutrition Examination Survey demographic data
}
\description{
This is a national survey run by CDC from 2015-16. More information at: https://wwwn.cdc.gov/Nchs/Nhanes/analyticguidelines.aspx
}
\usage{data("nhanesDemo")}
\format{
  A data frame with 9971 observations and 53 variables.
}

\source{
https://wwwn.cdc.gov/Nchs/Nhanes/2015-2016/DEMO_I.XPT
}

\examples{
data(nhanesDemo)
}
\keyword{datasets}
