\name{numHess1}
\alias{numHess1}
\docType{data}
\title{
numeric Hessian data 1
}
\description{
data file used by the HessianTest.R script
}
\usage{data("numHess1")}
\format{
  A 12 by 12 data frame containing Hessian (numeric variables a-l)
}
\examples{
data(numHess1)
str(numHess1)
}
\keyword{datasets}
