\name{numHess2}
\alias{numHess2}
\docType{data}
\title{
numeric Hessian data 2
}
\description{
data file used by the HessianTest.R script
}
\usage{data("numHess2")}
\format{
  A 12 by 12 data frame containing Hessian matrix (numeric variables a-l)
}
\examples{
data(numHess2)
str(numHess2)
}
\keyword{datasets}