% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxAutoStart.R
\name{omxBuildAutoStartModel}
\alias{omxBuildAutoStartModel}
\title{Build the model used for mxAutoStart}
\usage{
omxBuildAutoStartModel(model, type = c("ULS", "DWLS"))
}
\arguments{
\item{model}{The MxModel for which starting values are desired}

\item{type}{The type of starting values to obtain, currently unweighted or diagonally weighted least squares, ULS or DWLS}
}
\value{
an MxModel that can be run to obtain starting values
}
\description{
Build the model used for mxAutoStart
}
\seealso{
\link{mxAutoStart}
}
