% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxNamespace.R
\name{omxCheckNamespace}
\alias{omxCheckNamespace}
\title{omxCheckNamespace}
\usage{
omxCheckNamespace(model, namespace)
}
\arguments{
\item{model}{model}

\item{namespace}{namespace}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
\details{
This function checks that the named entities in the model are valid.
}
