% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxDetectCores.R
\name{omxDetectCores}
\alias{omxDetectCores}
\title{omxDetectCores}
\usage{
omxDetectCores(...)
}
\arguments{
\item{...}{unused}
}
\description{
Detects the number of cores on the local machine
}
\examples{
omxDetectCores()
}
