% Generated by roxygen2 (4.0.2): do not edit by hand
\name{omxGetNPSOL}
\alias{omxGetNPSOL}
\title{omxGetNPSOL}
\usage{
omxGetNPSOL()
}
\value{
Invisible NULL
}
\description{
Get the non-CRAN version of OpenMx from the OpenMx website.
}
\details{
This function
}
\examples{
\dontrun{omxGetNPSOL()}
}

