% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompute.R
\name{omxHasDefaultComputePlan}
\alias{omxHasDefaultComputePlan}
\title{omxHasDefaultComputePlan}
\usage{
omxHasDefaultComputePlan(model)
}
\arguments{
\item{model}{model}
}
\description{
Determine whether the model has a default complete plan (i.e., not custom).
}
