% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MxExpectationNormal.R
\name{omxManifestModelByParameterJacobian}
\alias{omxManifestModelByParameterJacobian}
\title{Estimate the Jacobian of manifest model with respect to parameters}
\usage{
omxManifestModelByParameterJacobian(model, defvar.row = 1)
}
\arguments{
\item{model}{an mxModel}

\item{defvar.row}{which row to use for definition variables}
}
\value{
a matrix with manifests in the rows and original parameters in the columns
}
\description{
The manifest model excludes any latent variables or processes. For
RAM and LISREL models, the manifest model contains only the
manifest variables with free means, covariance, and thresholds.
}
\details{
The Jacobian is estimated by the central finite difference.
}
\seealso{
\link{mxGetExpected}
}

