% Generated by roxygen2 (4.0.2): do not edit by hand
\name{omxNormalQuantiles}
\alias{omxNormalQuantiles}
\title{omxNormalQuantiles}
\usage{
omxNormalQuantiles(nBreaks, mean = 0, sd = 1)
}
\arguments{
\item{nBreaks}{the number of thresholds, or a vector of the number of thresholds}

\item{mean}{the mean of the underlying normal distribution}

\item{sd}{the standard deviation of the underlying normal distribution}
}
\value{
a vector of quantiles
}
\description{
Get quantiles from a normal distribution
}
\examples{
omxNormalQuantiles(3)
omxNormalQuantiles(3, mean=7)
omxNormalQuantiles(2, mean=1, sd=3)
}

