% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxRAMtoML.R
\name{omxRAMtoML}
\alias{omxRAMtoML}
\title{omxRAMtoML}
\usage{
omxRAMtoML(model)
}
\arguments{
\item{model}{the MxModel}
}
\value{
an ML model with an ML objective
}
\description{
Convert a RAM model to an ML model
}
\details{
This is a legacy function that was once used to convert RAM models to ML models
in the old (1.0 release of OpenMx) objective function style.
}
