\name{ordinalTwinData}
\alias{ordinalTwinData}
\docType{data}
\title{
Data for ordinal twin model
}
\description{
Example data for ordinal twin-data modelling. Three variables measured in each twin.
}
\usage{data("ordinalTwinData")}
\format{
  A data frame with 139 observations on the following 7 variables.
  \describe{
    \item{\code{zyg}}{a numeric vector}
    \item{\code{var1_twin1}}{a numeric vector}
    \item{\code{var2_twin1}}{a numeric vector}
    \item{\code{var3_twin1}}{a numeric vector}
    \item{\code{var1_twin2}}{a numeric vector}
    \item{\code{var2_twin2}}{a numeric vector}
    \item{\code{var3_twin2}}{a numeric vector}
  }
}
\examples{
data(ordinalTwinData)
str(ordinalTwinData)
}
\keyword{datasets}
