% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxFitFunction.R
\name{genericFitDependencies,MxBaseFitFunction-method}
\alias{genericFitDependencies,MxBaseFitFunction-method}
\title{Add dependencies}
\usage{
\S4method{genericFitDependencies}{MxBaseFitFunction}(.Object, flatModel, dependencies)
}
\arguments{
\item{.Object}{fit function object}

\item{flatModel}{flat model that lives with .Object}

\item{dependencies}{accumulated dependency relationships}
}
\description{
If there is an expectation, then the fitfunction should always
depend on it. Hence, subclasses that implement this method must
ignore the passed-in dependencies and use "dependencies <-
callNextMethod()" instead.
}
