% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxMatrix.R
\name{imxConDecMatrixSlots}
\alias{imxConDecMatrixSlots}
\alias{imxConDecMatrixSlots,MxMatrix-method}
\title{Condense/de-condense slots of an MxMatrix}
\usage{
imxConDecMatrixSlots(object)
}
\arguments{
\item{object}{of class MxMatrix}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
