% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxConstraint.R
\docType{data}
\name{imxConstraintRelations}
\alias{imxConstraintRelations}
\title{imxConstraintRelations}
\format{
An object of class \code{character} of length 3.
}
\usage{
imxConstraintRelations
}
\description{
A string vector of valid constraint binary relations.
}
\keyword{datasets}
