% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{imxConstraintRelations}
\alias{imxConstraintRelations}
\title{imxConstraintRelations}
\format{\preformatted{ chr [1:3] "<" "==" ">"
}}
\usage{
imxConstraintRelations
}
\description{
A string vector of valid constraint binary relations.
}
\keyword{datasets}

