% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxNamespace.R
\name{imxConvertIdentifier}
\alias{imxConvertIdentifier}
\title{imxConvertIdentifier}
\usage{
imxConvertIdentifier(identifiers, modelname, namespace, strict = FALSE)
}
\arguments{
\item{identifiers}{identifiers}

\item{modelname}{modelname}

\item{namespace}{namespace}

\item{strict}{strict}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
