% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxConvertIdentifier}
\alias{imxConvertIdentifier}
\title{imxConvertIdentifier}
\usage{
imxConvertIdentifier(identifiers, modelname, namespace)
}
\arguments{
\item{identifiers}{identifiers}

\item{modelname}{modelname}

\item{namespace}{namespace}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

