% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxConvertLabel}
\alias{imxConvertLabel}
\title{imxConvertLabel}
\usage{
imxConvertLabel(label, modelname, dataname, namespace)
}
\arguments{
\item{label}{label}

\item{modelname}{modelname}

\item{dataname}{dataname}

\item{namespace}{namespace}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

