% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxConvertSubstitution}
\alias{imxConvertSubstitution}
\title{imxConvertSubstitution}
\usage{
imxConvertSubstitution(substitution, modelname, namespace)
}
\arguments{
\item{substitution}{substitution}

\item{modelname}{modelname}

\item{namespace}{namespace}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

