% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxData.R
\docType{data}
\name{imxDataTypes}
\alias{imxDataTypes}
\title{Valid types of data that can be contained by MxData}
\format{
An object of class \code{character} of length 4.
}
\usage{
imxDataTypes
}
\description{
Valid types of data that can be contained by MxData
}
\keyword{datasets}
