% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{imxDataTypes}
\alias{imxDataTypes}
\title{Valid types of data that can be contained by MxData}
\format{\preformatted{ chr [1:5] "raw" "cov" "cor" "sscp" "acov"
}}
\usage{
imxDataTypes
}
\description{
Valid types of data that can be contained by MxData
}
\keyword{datasets}

