% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxDmvnorm}
\alias{imxDmvnorm}
\title{A C implementation of dmvnorm}
\usage{
imxDmvnorm(loc, mean, sigma)
}
\arguments{
\item{loc}{loc}

\item{mean}{mean}

\item{sigma}{sigma}
}
\description{
This API is visible to permit testing. Please do not use.
}

