% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxEvalByName}
\alias{imxEvalByName}
\title{imxEvalByName}
\usage{
imxEvalByName(name, model, compute = FALSE, show = FALSE)
}
\arguments{
\item{name}{name}

\item{model}{model}

\item{compute}{compute}

\item{show}{show}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

