% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxMultiModel.R
\name{imxFlattenModel}
\alias{imxFlattenModel}
\title{Remove hierarchical structure from model}
\usage{
imxFlattenModel(model, namespace, unsafe = FALSE)
}
\arguments{
\item{model}{model}

\item{namespace}{namespace}

\item{unsafe}{whether to skip sanity checks}
}
\description{
Remove hierarchical structure from model
}
