% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxFlattenModel}
\alias{imxFlattenModel}
\title{Remove heirarchical structure from model}
\usage{
imxFlattenModel(model, namespace)
}
\arguments{
\item{model}{model}

\item{namespace}{namespace}
}
\description{
Remove heirarchical structure from model
}

