% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxGenerateNamespace}
\alias{imxGenerateNamespace}
\title{imxGenerateNamespace}
\usage{
imxGenerateNamespace(model)
}
\arguments{
\item{model}{model}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

