% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxGenericModelBuilder}
\alias{imxGenericModelBuilder}
\title{imxGenericModelBuilder}
\usage{
imxGenericModelBuilder(model, lst, name, manifestVars, latentVars, submodels,
  remove, independent)
}
\arguments{
\item{model}{model}

\item{lst}{lst}

\item{name}{name}

\item{manifestVars}{manifestVars}

\item{latentVars}{latentVars}

\item{submodels}{submodels}

\item{remove}{remove}

\item{independent}{independent}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

