% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxGenSwift}
\alias{imxGenSwift}
\title{imxGenSwift}
\usage{
imxGenSwift(tc, sites, sfile)
}
\arguments{
\item{tc}{tc}

\item{sites}{sites}

\item{sfile}{sfile}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

