% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxExpectation.R
\name{imxHasThresholds}
\alias{imxHasThresholds}
\title{imxHasThresholds}
\usage{
imxHasThresholds(model)
}
\arguments{
\item{model}{model}
}
\description{
This is an internal function exported for those people who know
what they are doing.  This function checks if a model (or its
submodels) has any thresholds.
}
