% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\name{imxLog}
\alias{imxLog}
\title{Test thread-safe output code}
\usage{
imxLog(str)
}
\arguments{
\item{str}{the character string to output}
}
\description{
This is the code that the backend uses to write diagnostic
information to standard error. This function should not be called
from R. We make it available only for testing.
}

