% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxTypes.R
\docType{data}
\name{imxModelTypes}
\alias{imxModelTypes}
\title{imxModelTypes}
\format{
An object of class \code{list} of length 3.
}
\usage{
imxModelTypes
}
\description{
A list of supported model types
}
\keyword{datasets}
