% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OriginalMx.R
\name{imxOriginalMx}
\alias{imxOriginalMx}
\title{Run an classic mx script}
\usage{
imxOriginalMx(mx.filename, output.directory)
}
\arguments{
\item{mx.filename}{Name of file containing the mx script.}

\item{output.directory}{Where to write mxo output from the script}
}
\value{
processed matrix output.
}
\description{
For this to work, classic mx must be installed, and callable from the command line.
}
\examples{
\dontrun{
output = imxOriginalMx(mx.filename = "power1.mx", "~/Desktop")
}
}
