% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxReservedNames.R
\docType{data}
\name{imxReservedNames}
\alias{imxReservedNames}
\title{imxReservedNames}
\format{
An object of class \code{character} of length 7.
}
\usage{
imxReservedNames
}
\description{
Vector of reserved names
}
\keyword{datasets}
