% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{imxReservedNames}
\alias{imxReservedNames}
\title{imxReservedNames}
\format{\preformatted{ chr [1:6] "data" "objective" "likelihood" "fitfunction" ...
}}
\usage{
imxReservedNames
}
\description{
Vector of reserved names
}
\keyword{datasets}

