% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxReverseIdentifier}
\alias{imxReverseIdentifier}
\title{imxReverseIdentifier}
\usage{
imxReverseIdentifier(model, name)
}
\arguments{
\item{model}{model}

\item{name}{name}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

