% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxNamespace.R
\docType{data}
\name{imxSeparatorChar}
\alias{imxSeparatorChar}
\title{imxSeparatorChar}
\format{
An object of class \code{character} of length 1.
}
\usage{
imxSeparatorChar
}
\description{
The character between the model name and the named entity inside
the model.
}
\keyword{datasets}
