% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{imxSeparatorChar}
\alias{imxSeparatorChar}
\title{imxSeparatorChar}
\format{\preformatted{ chr "."
}}
\usage{
imxSeparatorChar
}
\description{
The character between the model name and the named entity inside
the model.
}
\keyword{datasets}

