% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxExpectationRAM.R
\name{imxSimpleRAMPredicate}
\alias{imxSimpleRAMPredicate}
\title{imxSimpleRAMPredicate}
\usage{
imxSimpleRAMPredicate(model)
}
\arguments{
\item{model}{model}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
