% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxTypeName}
\alias{imxTypeName}
\alias{imxTypeName,MxLISRELModel-method}
\alias{imxTypeName,MxModel-method}
\alias{imxTypeName,MxRAMModel-method}
\title{imxTypeName}
\usage{
imxTypeName(model)
}
\arguments{
\item{model}{model}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

