% Generated by roxygen2 (4.0.2): do not edit by hand
\name{imxUntitledNumber}
\alias{imxUntitledNumber}
\title{imxUntitledNumber}
\usage{
imxUntitledNumber()
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
\details{
Increments the untitled number counter and returns its value
}

