% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxRAMModel.R
\docType{data}
\name{imxVariableTypes}
\alias{imxVariableTypes}
\title{imxVariableTypes}
\format{
An object of class \code{character} of length 2.
}
\usage{
imxVariableTypes
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
\details{
The acceptable variable types
}
\keyword{datasets}
