% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MxRAMModel.R
\docType{data}
\name{imxVariableTypes}
\alias{imxVariableTypes}
\title{imxVariableTypes}
\format{\preformatted{ chr(0) 
}}
\usage{
imxVariableTypes
}
\description{
This is an internal function exported for those people who know
what they are doing.
}
\details{
The acceptable variable types
}
\keyword{datasets}

