% Generated by roxygen2 (4.0.0): do not edit by hand
\name{imxVerifyName}
\alias{imxVerifyName}
\title{imxVerifyName}
\usage{
imxVerifyName(name, stackNumber)
}
\arguments{
\item{name}{name}

\item{stackNumber}{stackNumber}
}
\description{
This is an internal function exported for those people who know
what they are doing.
}

