% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxAlgebra.R
\docType{class}
\name{MxAlgebraFormula-class}
\alias{MxAlgebraFormula-class}
\alias{MxAlgebraFormula}
\alias{print,MxAlgebraFormula-method}
\alias{show,MxAlgebraFormula-method}
\title{MxAlgebraFormula}
\description{
This is an internal class for the formulas used in \code{\link{mxAlgebra}} calls.
}
