% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxAvailableOptimizers.R
\name{mxAvailableOptimizers}
\alias{mxAvailableOptimizers}
\title{mxAvailableOptimizers}
\usage{
mxAvailableOptimizers()
}
\value{
- list of valid Optimizer names
}
\description{
List the Optimizers available in this version, e.g. "SLSQP" "CSOLNP"
}
\details{
note for advanced users: Special-purpose optimizers like Newton-Raphson or EM are not included in this list.
}
\examples{
mxAvailableOptimizers()
}
\seealso{
- \code{\link{mxOption}}(model, "Default optimizer")
}
\concept{Miscellaneous Utility Functions}
