% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxBaseNamed.R
\docType{class}
\name{MxBaseNamed-class}
\alias{MxBaseNamed-class}
\alias{MxBaseNamed}
\title{MxBaseNamed}
\description{
This is an internal class and should not be used directly.  It is the
base class for named entities.  Fit functions, expectations, and computes
contain this class.
}
