% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompute.R
\name{mxComputeDefault}
\alias{mxComputeDefault}
\alias{MxComputeDefault-class}
\title{Default compute plan}
\usage{
mxComputeDefault(freeSet = NA_character_)
}
\arguments{
\item{freeSet}{names of matrices containing free variables}
}
\description{
This is an empty placeholder for the default compute plan.
To create an actual plan, use \link{omxDefaultComputePlan}.
}
