% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mxComputeEM}
\alias{MxComputeEM-class}
\alias{mxComputeEM}
\title{Fit a model using DLR's (1977) Expectation-Maximization (EM) algorithm}
\usage{
mxComputeEM(expectation, predict, mstep, observedFit = "fitfunction", ...,
  maxIter = 500L, tolerance = 1e-04, verbose = 0L,
  freeSet = NA_character_, accel = "ramsay1975",
  information = NA_character_, infoArgs = list())
}
\arguments{
\item{expectation}{a vector of expectation names}

\item{predict}{what to predict from the observed data (available options depend on the expectation)}

\item{mstep}{a compute plan to optimize the completed data model}

\item{observedFit}{the name of the observed data fit function (defaults to "fitfunction")}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{maxIter}{maximum number of iterations}

\item{tolerance}{maximum change in fit to judge convergence}

\item{verbose}{level of diagnostic output}

\item{freeSet}{names of matrices containing free variables}

\item{accel}{name of acceleration method (defaults to "ramsay1975")}

\item{information}{name of information matrix approximation method}

\item{infoArgs}{arguments to control the information matrix method}
}
\description{
The EM algorithm constitutes the following steps: Start with an
initial parameter vector. Predict the missing data to form a
completed data model. Optimize the completed data model to obtain
a new parameter vector. Repeat these steps until convergence
criteria are met.
}
\details{
This compute plan does not work with any and all expectations. It
requires a special kind of expectation that can predict its
missing data to create a completed data model.

The EM algorithm does not produce a parameter covariance matrixn
for standard errors. S-EM, an implementation of Meng & Rubin
(1991), is included.
}
\references{
Dempster, A. P., Laird, N. M., & Rubin, D. B. (1977). Maximum likelihood from
incomplete data via the EM algorithm. \emph{Journal of the Royal Statistical Society.
Series B (Methodological)}, 1–38.

Meng, X.-L. & Rubin, D. B. (1991). Using EM to obtain asymptotic variance-covariance
matrices: The SEM algorithm. \emph{Journal of the American Statistical Association,
86} (416), 899–909.

Ramsay, J. O. (1975). Solving implicit equations in psychometric data analysis.
\emph{Psychometrika, 40} (3), 337–360.
}

