% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mxComputeHessianQuality}
\alias{MxComputeHessianQuality-class}
\alias{mxComputeHessianQuality}
\title{Compute the quality of the Hessian}
\usage{
mxComputeHessianQuality(freeSet = NA_character_)
}
\arguments{
\item{freeSet}{names of matrices containing free variables}
}
\description{
Tests whether the Hessian is positive definite
(model$output$infoDefinite) and, if so, computes the condition
number (model$output$conditionNumber). See Luenberger & Ye (2008)
Second Order Test (p. 190) and Condition Number (p. 239).
}
\references{
Luenberger, D. G. & Ye, Y. (2008). Linear and nonlinear programming. Springer.
}

