% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mxComputeIterate}
\alias{MxComputeIterate-class}
\alias{mxComputeIterate}
\title{Repeatedly invoke a series of compute objects until change is less than tolerance}
\usage{
mxComputeIterate(steps, ..., maxIter = 500L, tolerance = 1e-04,
  verbose = 0L, freeSet = NA_character_)
}
\arguments{
\item{steps}{a list of compute objects}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{maxIter}{the maximum number of iterations}

\item{tolerance}{iterates until change is less than tolerance}

\item{verbose}{level of debugging output}

\item{freeSet}{Names of matrices containing free variables.}
}
\description{
One step (typically the last) must compute the fit or maxAbsChange.
}

