% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompute.R
\name{mxComputeLoadContext}
\alias{mxComputeLoadContext}
\alias{MxComputeLoadContext-class}
\title{Load contextual data to supplement checkpoint}
\usage{
mxComputeLoadContext(
  method = c("csv"),
  path = c(),
  column,
  ...,
  sep = " ",
  verbose = 0L,
  header = TRUE,
  col.names = NULL
)
}
\arguments{
\item{method}{name of the conduit used to load the columns.}

\item{path}{the path to the file containing the data}

\item{column}{a character vector. The column names to log.}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{sep}{the field separator character. Values on each line of the file are separated by this character.}

\item{verbose}{integer. Level of run-time diagnostic output. Set to zero to disable}

\item{header}{logical. Whether the first row contains column headers.}

\item{col.names}{character vector. Column names}
}
\description{
\lifecycle{experimental}
}
\details{
Currently, this only supports comma separated value format and no
row names. If \code{header=TRUE} and \code{col.names} are
provided, the \code{col.names} take precedence. If
\code{header=FALSE} and no \code{col.names} are provided then
the column names consist of the file name and column offset.

An \code{originalDataIsIndexOne} option is not offered. You'll need to
add an extra line at the start on your file if you wish to make
use of \code{originalDataIsIndexOne} in \code{mxComputeLoad*}.
}
\seealso{
\link{mxComputeCheckpoint}, \link{mxComputeLoadData}, \link{mxComputeLoadMatrix}
}
