% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mxComputeNewtonRaphson}
\alias{MxComputeNewtonRaphson-class}
\alias{mxComputeNewtonRaphson}
\title{Optimize parameters using the Newton-Raphson algorithm}
\usage{
mxComputeNewtonRaphson(freeSet = NA_character_, ...,
  fitfunction = "fitfunction", maxIter = 100L, tolerance = 1e-07,
  verbose = 0L)
}
\arguments{
\item{freeSet}{names of matrices containing free variables}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{fitfunction}{name of the fitfunction (defaults to 'fitfunction')}

\item{maxIter}{maximum number of iterations}

\item{tolerance}{optimization is considered converged when maximum absolute change in parameters is less than tolerance}

\item{verbose}{level of debugging output}
}
\description{
This optimizer requires analytic 1st and 2nd derivatives of the
fit function. Ramsay (1975) is used to speed convergence. Ramsay
can be differentially applied to different groups of parameters.
Comprehensive diagnostics are available by increasing the verbose
level.
}
\references{
Luenberger, D. G. & Ye, Y. (2008). \emph{Linear and nonlinear programming.} Springer.

Ramsay, J. O. (1975). Solving implicit equations in psychometric data analysis.
\emph{Psychometrika, 40}(3), 337-360.
}

