% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompute.R
\name{mxComputeReportDeriv}
\alias{mxComputeReportDeriv}
\alias{MxComputeReportDeriv-class}
\title{Report derivatives}
\usage{
mxComputeReportDeriv(freeSet = NA_character_)
}
\arguments{
\item{freeSet}{names of matrices containing free variables}
}
\description{
Copy the internal gradient and Hessian back to R.
}
