% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mxComputeReportDeriv}
\alias{MxComputeReportDeriv-class}
\alias{mxComputeReportDeriv}
\title{Report derivatives}
\usage{
mxComputeReportDeriv(freeSet = NA_character_)
}
\arguments{
\item{freeSet}{names of matrices containing free variables}
}
\description{
Copy the internal gradient and Hessian back to R.
}

