% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mxComputeSequence}
\alias{MxComputeSequence-class}
\alias{mxComputeSequence}
\title{Invoke a series of compute objects in sequence}
\usage{
mxComputeSequence(steps = list(), ..., freeSet = NA_character_)
}
\arguments{
\item{steps}{a list of compute objects}

\item{...}{Not used; forces argument 'freeSet' to be specified by name.}

\item{freeSet}{Names of matrices containing free parameters.}
}
\description{
Invoke a series of compute objects in sequence
}

