% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompute.R
\name{mxComputeStandardError}
\alias{mxComputeStandardError}
\alias{MxComputeStandardError-class}
\title{Compute standard errors}
\usage{
mxComputeStandardError(freeSet = NA_character_,
  fitfunction = "fitfunction")
}
\arguments{
\item{freeSet}{names of matrices containing free variables}

\item{fitfunction}{name of the fitfunction (defaults to 'fitfunction')}
}
\description{
When the fit is in -2 log likelihood units, the SEs are derived
from the diagonal of the Hessian or inverse Hessian. The Hessian
(in some form) must already be available.
}
