% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxCompute.R
\name{mxComputeStandardError}
\alias{MxComputeStandardError-class}
\alias{mxComputeStandardError}
\title{Compute standard errors given the Hessian or inverse Hessian}
\usage{
mxComputeStandardError(freeSet = NA_character_)
}
\arguments{
\item{freeSet}{names of matrices containing free variables}
}
\description{
The fit is assumed to be in deviance units (-2 log likelihood).
}

