% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mxComputeStandardError}
\alias{MxComputeStandardError-class}
\alias{mxComputeStandardError}
\title{Compute standard errors given the Hessian or inverse Hessian}
\usage{
mxComputeStandardError(freeSet = NA_character_)
}
\arguments{
\item{freeSet}{names of matrices containing free variables}
}
\description{
The fit is assumed to be in deviance units (-2 log likelihood).
}

