% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mxDataDynamic}
\alias{MxDataDynamic-class}
\alias{mxDataDynamic}
\title{Create dynamic data}
\usage{
mxDataDynamic(type, ..., expectation)
}
\arguments{
\item{type}{type of data}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{expectation}{the name of the expectation to provide the data}
}
\description{
Create dynamic data
}

