% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxData.R
\name{MxDataFrameOrMatrix-class}
\alias{MxDataFrameOrMatrix}
\alias{MxDataFrameOrMatrix-class}
\title{MxDataFrameOrMatrix}
\description{
Internal class that is the union of data.frame and matrix.
}
\details{
Not to be used.
}

