% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mxExpectationBA81}
\alias{mxExpectationBA81}
\title{Create a Bock & Aitkin (1981) expectation}
\usage{
mxExpectationBA81(ItemSpec, ItemParam, design = NULL, qpoints = NULL,
  qwidth = 6, mean = NULL, cov = NULL, scores = "omit", verbose = 0L,
  EItemParam = NULL, debugInternal = FALSE)
}
\arguments{
\item{ItemParam}{one column for each item with parameters starting
at row 1 and extra rows filled with NA}

\item{design}{one column per item, assignment of person abilities
to item dimensions (optional)}

\item{qpoints}{number of points to use for rectangular quadrature integrations (default 49)
See Seong (1990) for some considerations on specifying this parameter.}
}
\description{
The standard Normal distribution of the quadrature acts like a
prior distribution for difficulty. It is not necessary to impose
any additional Bayesian prior on difficulty estimates (Baker &
Kim, 2004, p. 196).
}
\references{
Bock, R. D., & Aitkin, M. (1981). Marginal maximum likelihood estimation of item
parameters: Application of an EM algorithm. Psychometrika, 46, 443-459.

Cai, L. (2010). A two-tier full-information item factor analysis
model with applications. Psychometrika, 75, 581-612.

Seong, T. J. (1990). Sensitivity of marginal maximum likelihood
estimation of item and ability parameters to the characteristics
of the prior ability distributions. Applied Psychological
Measurement, 14(3), 299-311.
}

