% Generated by roxygen2 (4.0.0): do not edit by hand
\name{mxFitFunctionMultigroup}
\alias{MxFitFunctionMultigroup-class}
\alias{mxFitFunctionMultigroup}
\title{Aggregate fit statistics from submodels}
\usage{
mxFitFunctionMultigroup(groups, ..., verbose = 0L)
}
\arguments{
\item{groups}{vector of fit function names}

\item{...}{Not used.  Forces remaining arguments to be specified by name.}

\item{verbose}{the level of debugging output}
}
\description{
This is a simple fit function that sums the fit statistics
from other fit functions, typically in submodels. It is almost
equivalent to,
}
\details{
\code{mxAlgebra(model1.objective + model2.objective, name="alg")}

and

\code{mxFitFunctionAlgebra("alg")}

However, in addition to the fit statistic, mxFitFunctionMultigroup
also aggregates analytic derivative calculations.
}

