%
%   Copyright 2007-2014 The OpenMx Project
%
%   Licensed under the Apache License, Version 2.0 (the "License");
%   you may not use this file except in compliance with the License.
%   You may obtain a copy of the License at
% 
%        http://www.apache.org/licenses/LICENSE-2.0
% 
%   Unless required by applicable law or agreed to in writing, software
%   distributed under the License is distributed on an "AS IS" BASIS,
%   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
%   See the License for the specific language governing permissions and
%   limitations under the License.

\name{mxFitFunctionMultigroup}
\alias{MxFitFunctionMultigroup-class}
\alias{mxFitFunctionMultigroup}

\title{Aggregate fit statistics from submodels}

\description{
This is fit function is used to sum the fit statistics
from other fit functions, typically in submodels.

This call:

    \code{mxFitFunctionMultigroup(c("model1", "model2"))}

is almost equivalent to the following pair of statements:

    \code{mxAlgebra(model1.objective + model2.objective, name="alg")}

    \code{mxFitFunctionAlgebra("alg")}

However, in addition to the fit statistic, mxFitFunctionMultigroup
also aggregates analytic derivative calculations.

A further advantage is that this allows \code{\link{mxSaturatedModel}} 
compute saturated models for raw data.

Note: You can refer to the algebra generated by mxFitFunctionMultigroup 
as \code{modelName.fitfunction}

}

\usage{
mxFitFunctionMultigroup(groups, ..., verbose = 0L)
}

\arguments{
\item{groups}{vector of fit function names}

\item{...}{Not used. Forces remaining arguments to be specified by name.}

\item{verbose}{the level of debugging output}
}

\examples{
require("OpenMx")
mxFitFunctionMultigroup(c("model1", "model2")) # names of sub-models to be jointly optimised
}
