% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MxInterval.R
\docType{class}
\name{MxInterval-class}
\alias{$,MxInterval-method}
\alias{$<-,MxInterval-method}
\alias{MxInterval-class}
\alias{print,MxInterval-method}
\alias{show,MxInterval-method}
\title{MxInterval}
\description{
This is an internal class and should not be used directly.
}
\seealso{
\link{mxCI}
}

