\name{mxVersion}
\alias{mxVersion}

\title{Returns Current Version String}

\description{
   This function returns a string with the current version number of OpenMx.
}

\usage{
mxVersion()
}

\references{
The OpenMx User's guide can be found at http://openmx.psyc.virginia.edu/documentation.
}

\examples{

mxVersion()

}
