% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omxConstrainThresholds.R
\name{omxConstrainMLThresholds}
\alias{omxConstrainMLThresholds}
\title{omxConstrainMLThresholds}
\usage{
omxConstrainMLThresholds(model, dist = 0.1)
}
\arguments{
\item{model}{the MxModel to which constraints should be added}

\item{dist}{unused}
}
\value{
a new MxModel object with the constraints added
}
\description{
Add constraint to ML model to keep thresholds in order
}
\details{
This function adds a nonlinear constraint to an ML model.  The constraint
keeps the thresholds in order.  Constraints often slow model estimation,
however, keeping the thresholds in increasing order helps ensure the likelihood
function is well-defined.  If you're having problems with ordinal data, this is
one of the things to try.
}
\seealso{
\code{demo("omxConstrainMLThresholds")}
}
