% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omxGetNPSOL.R
\name{omxGetNPSOL}
\alias{omxGetNPSOL}
\title{omxGetNPSOL}
\usage{
omxGetNPSOL()
}
\value{
Invisible NULL
}
\description{
Get the non-CRAN version of OpenMx from the OpenMx website.
}
\details{
This function
}
