% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MxModelParameters.R
\name{omxNameAnonymousParameters}
\alias{omxNameAnonymousParameters}
\title{omxNameAnonymousParameters}
\usage{
omxNameAnonymousParameters(model, indep = FALSE)
}
\arguments{
\item{model}{the MxModel}

\item{indep}{whether models are independent}
}
\value{
a list with components for the new MxModel with named parameters, and the new names.
}
\description{
Assign new names to the unnamed parameters
}
