\name{OpenMx}
\alias{OpenMx}

\title{OpenMx: Package for Matrix Algebra Optimization}

\description{OpenMx is a package for structural equation modeling, matrix algebra optimization and other statistical estimation problems. Objects used or created by OpenMx may be of the following classes: \link{MxAlgebra}, \link{MxBounds}, \link{MxConstraint}, \link{MxData}, \link{MxMatrix} and \link{MxModel}. Objects of these classes may be created by the following OpenMx functions: \link{mxAlgebra}, \link{mxBounds}, \link{mxConstraint}, \link{mxData}, \link{mxMatrix} and \link{mxModel}.

}

\details{
Details coming soon.
}

\references{
For more information about OpenMx, 
}